imma party first

+++ Util/Comparator/Minutiae.cs
@@ -3,14 +3,14 @@ namespace FingerprintRecognitionV2.Util.Comparator
 {
     public class Minutiae
     {
-        static public readonly int NO_TYPE = 0, ENDING = 1, BIFUR = 2;
+        static public readonly byte NO_TYPE = 0, ENDING = 1, BIFUR = 2;
 
-        public double T;    // type
+        public byte T;      // type
         public double Y;    // y location
         public double X;    // x location
         public double A;    // the angle
 
-        public Minutiae(int t, double y, double x, double a)
+        public Minutiae(byte t, double y, double x, double a)
         {
             T = t;
             Y = y;
